fix apis

+++ src/lib/api/services/device.ts
@@ -1,16 +1,17 @@
+import { Client } from '@/lib/api/client'
 import { ApiResponse } from '@/types/api/common'
 import { Device } from '@/types/user'
 
-import { Client } from '../client'
-
 export const createDeviceService = (client: Client) => ({
-  index: async (): Promise<Device[]> => {
-    const response = await client.get<ApiResponse<Device[]>>(`/api/devices`)
+  index: async (userId: number): Promise<Device[]> => {
+    const response = await client.get<ApiResponse<Device[]>>(
+      `/api/users/${userId}/devices`
+    )
     return response.data
   },
 
-  destroy: async (uuid: string): Promise<void> => {
-    await client.delete(`/api/devices/${uuid}`)
+  destroy: async (userId: number, uuid: string): Promise<void> => {
+    await client.delete(`/api/users/${userId}/devices/${uuid}`)
   },
 })
